refactor: 나만의 기준으로 폴더 구조 정리
