+++ my-app/tsconfig.json
@@ -24,7 +24,7 @@
   "include": [
     "src"
   ],
-  "baseUrl": "./", // 기본 경로 설정
+  "baseUrl": "src", // 기본 경로 설정
   "paths": { //절대 경로를 지정하고 싶은 경로들을 지정
     "reset-css": ["types/reset-css"]
   }
